move test cases plus setup/teardown to subclasses
